+++ noodles-gff/src/async/io/reader.rs
@@ -1,9 +1,8 @@
-mod lazy_line;
+mod line;
 
 use futures::{stream, Stream, TryStreamExt};
 use tokio::io::{self, AsyncBufRead, AsyncBufReadExt};
 
-use self::lazy_line::read_lazy_line;
 use crate::{DirectiveBuf, Line, LineBuf, RecordBuf};
 
 /// An async GFF reader.
@@ -108,15 +107,15 @@ where
     ///
     /// let mut line = gff::Line::default();
     ///
-    /// reader.read_lazy_line(&mut line).await?;
+    /// reader.read_line(&mut line).await?;
     /// assert_eq!(line.kind(), gff::line::Kind::Directive);
     ///
-    /// assert_eq!(reader.read_lazy_line(&mut line).await?, 0);
+    /// assert_eq!(reader.read_line(&mut line).await?, 0);
     /// # Ok(())
     /// # }
     /// ```
-    pub async fn read_lazy_line(&mut self, line: &mut Line) -> io::Result<usize> {
-        read_lazy_line(&mut self.inner, line).await
+    pub async fn read_line(&mut self, line: &mut Line) -> io::Result<usize> {
+        line::read_line(&mut self.inner, line).await
     }
 
     /// Returns a stream over line buffers.